Fix some small issues
This commit fixes some small issues with various parts of the
deployment. This is mostly just missing dependencies in the images'
Dockerfiles and a couple of syntax errors. The real change of note
though is the initContainer used for DB migrations is switched to have
the command in the k8s yaml instead of the image build.