Fix minor oversights in api initContainer
The api init container is setup to both wait for the posstgres db to be
running and also ensure the db schema is setup properly. There were some
issues in the script it was supposed to run however. This commit fixes
those issues. However, the init container is still failing to execute.