Fix issues with API deployment
This commit fixes several issues with the API deployment which
prevented the API server from running. After this commit the API server
will properly launch and respond to API requests. There is still an
issue with the setup and the database information is not getting
injected into the openstack-health api server config file.