Fix DB setup for o-h api server
This commit fixes the DB configuration for the o-h server. Previously we
weren't setting up the env variables correctly for the deployment which
caused an invalid db_uri to be written into the confi file. This commit
fixes that and by extension fixes are secrets export from the init
contianer of the DB host address so that correct host gets set in the
config file.