Add ingress and template the skaffold release name
For CI, each PR should deploy to a different release name at least.
This depends on https://github.com/GoogleContainerTools/skaffold/pull/602
which is not merged yet in skaffold.

Add ingress as well, so that the cluster deployed via CI can be accessed
via a public URL. For simplicity for now I assume everything is in the
default namespace and so to avoid conflict between releases I defined
a release specific path.

The ingress configuration are a simplified version of what "helm create"
would create. Eventually it would be good to move annotations and paths
back into values to make it possible to run this chart on a ingress
implementation other then the IKS custom one.

In future we could use per-ci-build (i.e. per-release) sub-domains and
a fixed path to access api and dashboard, but that requires some extra
work especially if we want to enable TLS as well (which we should).

See https://console.bluemix.net/docs/containers/cs_ingress.html#ingress