Add initial travis configuration
Use a VM node with docker to build images.
Use a go image to build skaffold with PR 602.

Only deploy to a cluster in case of pull request. This is just an
initial stub at this, the idea is to build/update the cluster when
the PR is created/updated and to delete it when it's merged.

For a push commit I'd create the cluster, run integration tests
and delete it right away. For other PR actions like delete, merge,
comment, etc I'd use a bot to trigger cleanup or update of the
clsuter.

The CI job runs fine to the helm deployment however it often fails
now as it hits a race condition in the pods creation.