Test both push and PR
Deploy to a test cluster with a release named after the branch.
If it's a push, wait for the deploy to be done and then cleanup.
If it's a PR, wait for the deploy to be done and, in case of
succcessful deployment, leave the deploy around.

We may want in future to leave the cluster around in case of
failure as well, but we need to change skaffold to be able to
deal with a failed deploy. At the moment it will try upgrade
which won't work. Or to make the CI script handle that.