Switch version tag to string
This commit switches version string in the values.yaml to be a string
instead of a int. The template processing for the images was expecting
this to be a string and was failing in the templating by exposing the
float type instead of the raw string. This works around that by just
being explicit that we me the number as a string instead of relying on
the templating to do the right thing.