Drop the api-wait image
Instead of waiting for API to be up and running and get the LB
IP address, use the ingress name directly.
Move the protocol part into the envirnment variable so we can
control it when we enabled (optional) TLS.

The case of ingress disabled is not really handled yet.
We should use the service name there, but the client would need
to be able to resolve it.