Add junitxml2subunit script
This commit adds a the junitxml2subunit script which reads a junitxml
results file and converts that into subunit v2. The subunit protocol
expects a bit more information then what's in junitxml, specifically
around timestamps. To fake this utility uses the current time when
running the command and increases that by the duration of each test
for each consecutive test.